+++ pkg/tuple/tuple.go
@@ -63,21 +63,6 @@ func GetObjectRelationAsString(objectRelation *openfgapb.ObjectRelation) string 
 	return objectRelation.GetObject()
 }
 
-// GetRelationReferenceAsString returns team#member, or team:*, or an empty string if the input is nil.
-func GetRelationReferenceAsString(rr *openfgapb.RelationReference) string {
-	if rr == nil {
-		return ""
-	}
-	if _, ok := rr.RelationOrWildcard.(*openfgapb.RelationReference_Relation); ok {
-		return fmt.Sprintf("%s#%s", rr.GetType(), rr.GetRelation())
-	}
-	if _, ok := rr.RelationOrWildcard.(*openfgapb.RelationReference_Wildcard); ok {
-		return fmt.Sprintf("%s:*", rr.GetType())
-	}
-
-	panic("unexpected relation reference")
-}
-
 // SplitObjectRelation splits an object relation string into an object ID and relation name. If no relation is present,
 // it returns the original string and an empty relation.
 func SplitObjectRelation(objectRelation string) (string, string) {
